wire.repeater_fraction now always specified from lower to higher metal layers -- TDWLARI now flips it
Now the first element of wire.repeater_fraction always corresponds to a
lower metal layer than the top layer. If TDWLARI is not enabled, the
first element is the 1st metal layer. if TDWLARI is enabled, the last
element is the top metal layer, and the first element is the top-nth
metal layer (TDWLARI does fliplr(wire.repeater_fraction))

-Updated all the test cases so they still work as intended