[WLA] Alternate metals can now be used for wires smaller than a minimum pitch to mitigate electromigration concerns
wire.use_em_resistant_metal enables or disables this behavior
wire.min_non_em_width sets the minimum width for which copper can be
used
wire.alt_resistivity_em is the bulk resistivity of the alternate
material

For now all alternate materials are still using the same MS+FS model as
the copper wires, even with the same specularity and reflectivity
coefficients. This is due to the lack of data for many of these
materials. Eventually the specularity and reflectivity will be
user-settable, rather than hardcoded in WLATDRI