[POWER] [RE] Removed Vdd and thermal impacts on leakage due to incorrect leakage values
the previous way of doing things was giving erroneously low values. This
is partly due to the fact that Ioff already accounts for typical
Vdd/Vgs/Vth conditions, since it's such a coarse measure.

The leakage calculation must be revisited in the future to improve its
accuracy and restore the thermal dependence