WLATDRI now attempts to partially route wires of each length until area limit is reached for each tier
Previously It would only route ALL wires of a particular length at once,
and if it couldn't route all of them it would just not route any. This
way we get better fill factors, and potentially avoid getting stuck if
we can't route all the wires on a particular tier.